+++ write_config_files/commands/write.py
@@ -11,6 +11,7 @@ from ..rendering import TemplateRenderer
 
 class FileSystem(Protocol):
     def file_exists(self, path: str) -> bool: ...
+    def read_content(self, path: str) -> str: ...
     def write_file(self, path: str, content: str, is_executable: bool) -> None: ...
 
 
@@ -22,9 +23,10 @@ class Writer:
 
     def write(self, templates: TemplateConfig, skip_if_exists: bool) -> None:
         for file in templates.files:
-            rendered = self.renderer.render(file.template_name)
+            rendered_content = self.renderer.render(file.template_name)
 
             already_exists = self.file_system.file_exists(file.destination_path)
+            current_content = self.file_system.read_content(file.destination_path)
 
             if not already_exists:
                 self.logger.info(f'writing {file.destination_path}')
@@ -33,11 +35,16 @@ class Writer:
                     f'skipping {file.destination_path} because it already exists',
                 )
                 continue
+            elif rendered_content == current_content:
+                self.logger.debug(
+                    f'skipping {file.destination_path} because there are no changes',
+                )
+                continue
             else:
                 self.logger.warn(f'overwriting {file.destination_path}')
 
             self.file_system.write_file(
                 file.destination_path,
-                rendered,
+                rendered_content,
                 file.is_executable,
             )